+++ android/quest/src/main/java/org/smartregister/fhircore/quest/data/patient/model/DetailsModel.kt
@@ -14,13 +14,19 @@
  * limitations under the License.
  */
 
-package org.smartregister.fhircore.quest.util
+package org.smartregister.fhircore.quest.data.patient.model
 
-import org.smartregister.fhircore.engine.configuration.ConfigClassification
+import androidx.compose.runtime.Stable
+import org.hl7.fhir.r4.model.Type
+import org.smartregister.fhircore.quest.configuration.view.Filter
 
-enum class QuestConfigClassification : ConfigClassification {
-  PATIENT_REGISTER,
-  PATIENT_REGISTER_ROW,
-  TEST_RESULT_DETAIL_VIEW;
-  override val classification: String = name.lowercase()
-}
+@Stable
+data class DetailsViewItem(
+  var label: String = "",
+  val rows: MutableList<DetailsViewItemRow> = mutableListOf()
+)
+
+@Stable
+data class DetailsViewItemRow(val cells: MutableList<DetailsViewItemCell> = mutableListOf())
+
+@Stable data class DetailsViewItemCell(val value: Type?, val filter: Filter)